+++ internal/cmlschema/group_test.go
@@ -21,13 +21,10 @@ var group1 *cmlclient.Group = &cmlclient.Group{
 		"90f84e38-a71c-4d57-8d90-00fa8a197385",
 		"60f84e39-ffff-4d99-8a78-00fa8aaf5666",
 	},
-	Labs: []struct {
-		ID         string `json:"id"`
-		Permission string `json:"permission"`
-	}{
+	Labs: []cmlclient.GroupLab{
 		{
-			"90f84e38-a71c-4d57-8d90-00fa8a197385",
-			"read_only",
+			ID:         "90f84e38-a71c-4d57-8d90-00fa8a197385",
+			Permission: "read_only",
 		},
 	},
 	ID: "85401911-851f-4e6a-b5c3-4aa1d91fa21d",
@@ -37,11 +34,8 @@ var group2 *cmlclient.Group = &cmlclient.Group{
 	Name:        "CCNA Study Group Class of 01",
 	Description: "string",
 	Members:     []string{},
-	Labs: []struct {
-		ID         string `json:"id"`
-		Permission string `json:"permission"`
-	}{},
-	ID: "85401911-851f-4e6a-b5c3-4aa1d91fa21d",
+	Labs:        []cmlclient.GroupLab{},
+	ID:          "85401911-851f-4e6a-b5c3-4aa1d91fa21d",
 }
 
 func TestGroup(t *testing.T) {