+++ internal/provider/resource/group/meta.go
@@ -13,7 +13,6 @@ import (
 // Ensure provider defined types fully satisfy framework interfaces
 var _ resource.Resource = &GroupResource{}
 var _ resource.ResourceWithImportState = &GroupResource{}
-var _ resource.ResourceWithModifyPlan = &GroupResource{}
 
 type GroupResource struct {
 	cfg *common.ProviderConfig