+++ internal/provider/resource/user/read.go
@@ -1,4 +1,4 @@
-package group
+package user
 
 import (
 	"context"
@@ -10,10 +10,10 @@ import (
 	"github.com/rschmied/terraform-provider-cml2/internal/common"
 )
 
-func (r *GroupResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
-	var data cmlschema.GroupModel
+func (r *UserResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	var data cmlschema.UserModel
 
-	tflog.Info(ctx, "Resource Group READ")
+	tflog.Info(ctx, "Resource User READ")
 
 	// Read Terraform configuration data into the model
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
@@ -21,19 +21,21 @@ func (r *GroupResource) Read(ctx context.Context, req resource.ReadRequest, resp
 		return
 	}
 
-	group, err := r.cfg.Client().GroupGet(ctx, data.ID.ValueString())
+	user, err := r.cfg.Client().UserGet(ctx, data.ID.ValueString())
 	if err != nil {
 		resp.Diagnostics.AddError(
 			common.ErrorLabel,
-			fmt.Sprintf("Unable to get group, got error: %s", err),
+			fmt.Sprintf("Unable to get user, got error: %s", err),
 		)
 		return
 	}
 
-	value := cmlschema.NewGroup(ctx, group, &resp.Diagnostics)
+	// need to preserve "write once" values
+	user.Password = data.Password.ValueString()
+	value := cmlschema.NewUser(ctx, user, &resp.Diagnostics)
 
 	// Save data into Terraform state
 	resp.Diagnostics.Append(resp.State.Set(ctx, &value)...)
 
-	tflog.Info(ctx, "Resource Group READ: done")
+	tflog.Info(ctx, "Resource User READ: done")
 }